+++ proxi-alert-engine/src/main/java/code/adagedo/proxialertengine/repositories/UserRepository.java
@@ -4,12 +4,13 @@ import code.adagedo.proxialertengine.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     List<User> findByLatitudeBetweenAndLongitudeBetween(
-            Double minLat, Double maxLat,
-            Double minLon, Double maxLon
+            BigDecimal minLat, BigDecimal maxLat,
+            BigDecimal minLon, BigDecimal maxLon
     );
 }
